WIP Ensure autorequire is case insensitive
Developing the acceptance tests for the registry_value, I noticed
autorequire doesn't work if the parent registry_key resource doesn't
match the same same case as the registry_value being managed.

This is a problem because the windows ecosystem is case insensitive and
case preserving.

This patch adds test coverage to ensure a key path expressed in upper
case is automatically required by a value path expressed in lower case.